+++ apps/prairielearn/src/lib/client/safe-db-types.ts
@@ -183,6 +183,8 @@ export const RawStudentCourseInstanceSchema = RawStaffCourseInstanceSchema.pick(
   hide_in_enroll_page: true,
   id: true,
   long_name: true,
+  publishing_end_date: true,
+  publishing_start_date: true,
   short_name: true,
 });
 export const StudentCourseInstanceSchema =
@@ -221,6 +223,17 @@ export const StudentEnrollmentSchema = RawStudentEnrollmentSchema.brand<'Student
 export type StudentEnrollment = z.infer<typeof StudentEnrollmentSchema>;
 
 /** Institutions */
+export const RawAdminInstitutionSchema = RawInstitutionSchema.pick({
+  default_authn_provider_id: true,
+  display_timezone: true,
+  id: true,
+  long_name: true,
+  short_name: true,
+  uid_regexp: true,
+});
+export const AdminInstitutionSchema = RawAdminInstitutionSchema.brand<'AdminInstitution'>();
+export type AdminInstitution = z.infer<typeof AdminInstitutionSchema>;
+
 export const RawStaffInstitutionSchema = RawInstitutionSchema.pick({
   default_authn_provider_id: true,
   display_timezone: true,
@@ -252,8 +265,7 @@ export const StaffTagSchema = RawTagSchema.brand<'StaffTag'>();
 export type StaffTag = z.infer<typeof StaffTagSchema>;
 
 /** Users */
-
-const RawStaffUserSchema = RawUserSchema.pick({
+export const RawStaffUserSchema = RawUserSchema.pick({
   email: true,
   institution_id: true,
   name: true,
@@ -261,6 +273,7 @@ const RawStaffUserSchema = RawUserSchema.pick({
   uin: true,
   user_id: true,
 });
+export type RawStaffUser = z.infer<typeof RawStaffUserSchema>;
 export const StaffUserSchema = RawStaffUserSchema.brand<'StaffUser'>();
 export type StaffUser = z.infer<typeof StaffUserSchema>;