+++ apps/prairielearn/src/pages/instructorQuestionSettings/instructorQuestionSettings.html.tsx
@@ -13,7 +13,6 @@ import { TagDescription } from '../../components/TagDescription.js';
 import { TopicBadgeHtml } from '../../components/TopicBadge.js';
 import { TopicDescription } from '../../components/TopicDescription.js';
 import { compiledScriptTag, nodeModulesAssetPath } from '../../lib/assets.js';
-import { config } from '../../lib/config.js';
 import {
   AssessmentSchema,
   AssessmentSetSchema,
@@ -228,7 +227,7 @@ export function InstructorQuestionSettings({
                                         data-name="${tag.name}"
                                         data-description="${tag.implicit
                                           ? ''
-                                          : TagDescription(tag)}"
+                                          : renderHtml(<TagDescription tag={tag} />)}"
                                         ${questionTagNames.has(tag.name) ? 'selected' : ''}
                                       ></option>
                                     `;
@@ -766,7 +765,6 @@ function DeleteQuestionModal({
                     <div class="h6">${a_with_q.short_name} (${a_with_q.long_name})</div>
                     ${a_with_q.assessments.map((assessment) =>
                       AssessmentBadgeHtml({
-                        plainUrlPrefix: config.urlPrefix,
                         courseInstanceId: a_with_q.course_instance_id,
                         assessment,
                       }),