+++ apps/prairielearn/src/tests/exampleCourseQuestionsComplete.test.ts
@@ -92,7 +92,7 @@ const rewriteValidatorFalsePositives = async (html: string): Promise<string> => 
 
 const validateHtml = async (html: string) => {
   const rewrittenHtml = await rewriteValidatorFalsePositives(html);
-  const { valid, results } = await htmlvalidate.validateString(rewrittenHtml, {
+  const { results } = await htmlvalidate.validateString(rewrittenHtml, {
     extends: ['html-validate:recommended', 'html-validate:document'],
     rules: {
       // https://html-validate.org/rules/no-raw-characters.html
@@ -177,11 +177,25 @@ const validateHtml = async (html: string) => {
       'element-permitted-content': 'off',
     },
   });
+
+  const filteredResults = results.map((result) => {
+    result.messages = result.messages.filter((m) => {
+      // Workaround for https://gitlab.com/html-validate/html-validate/-/issues/334
+      if (m.ruleId === 'aria-label-misuse' && m.selector && /^.*> option[^>]*$/.test(m.selector)) {
+        return false;
+      }
+      return true;
+    });
+    return result;
+  });
+
+  const valid = filteredResults.every((result) => result.messages.length === 0);
+
   if (!valid) {
-    const validationMessages = results.flatMap((result) =>
+    const validationMessages = filteredResults.flatMap((result) =>
       result.messages.map((m) => `L${m.line}:C${m.column} ${m.message} (${m.ruleId})`),
     );
-    assert.fail(`HTMLValidate failed:\n${validationMessages.join('\n')}\n${rewrittenHtml}`);
+    assert.fail(`HTMLValidate failed:\n${rewrittenHtml}\n${validationMessages.join('\n')}`);
   }
 };
 
@@ -294,7 +308,7 @@ describe('Internally graded question lifecycle tests', { timeout: 60_000 }, func
       // Render
       const locals = {
         urlPrefix: '/prefix1',
-        plainUrlPrefix: config.urlPrefix,
+        plainUrlPrefix: '/pl',
         questionRenderContext: undefined,
         ...buildQuestionUrls(
           '/prefix2',