+++ packages/utils/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @prairielearn/utils
 
+## 2.0.3
+
+### Patch Changes
+
+- 0425922: Upgrade all JavaScript dependencies
+
+## 2.0.2
+
+### Patch Changes
+
+- c0b1c74: Enable `declarationMap`
+
 ## 2.0.1
 
 ### Patch Changes